+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added transaction tracker helpers tests](https://github.com/multiversx/mx-sdk-dapp/pull/1627)
 - [Fixed explorer link console error](https://github.com/multiversx/mx-sdk-dapp/pull/1626)
 - [Added NotificationsFeedManager basic tests](https://github.com/multiversx/mx-sdk-dapp/pull/1625)
 - [Fixed NotificationsFeedManager hanging toasts](https://github.com/multiversx/mx-sdk-dapp/pull/1622)